luis-scripts/requirements.txt
adal==1.2.7
azure-cognitiveservices-language-luis==0.7.0
azure-common==1.1.28
azure-core==1.26.0
black==22.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.2
cryptography==38.0.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
msrestazure==0.6.4
mypy-extensions==0.4.3
oauthlib==3.2.1
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
